Rename config files to match current coding standards
