+++ cumulusci/core/config/__init__.py
@@ -3,7 +3,7 @@
 # constants used by MetaCI
 FAILED_TO_CREATE_SCRATCH_ORG = "Failed to create scratch org"
 
-from cumulusci.core.config.BaseConfig import BaseConfig
+from cumulusci.core.config.base_config import BaseConfig
 
 # inherit from BaseConfig
 
@@ -20,7 +20,7 @@ class FlowConfig(BaseConfig):
     pass
 
 
-from cumulusci.core.config.OrgConfig import OrgConfig
+from cumulusci.core.config.org_config import OrgConfig
 
 
 class ServiceConfig(BaseConfig):
@@ -35,21 +35,19 @@ class TaskConfig(BaseConfig):
     pass
 
 
-from cumulusci.core.config.BaseTaskFlowConfig import BaseTaskFlowConfig
-
+from cumulusci.core.config.base_task_flow_config import BaseTaskFlowConfig
 
 # inherit from BaseTaskFlowConfig
 from cumulusci.core.config.project_config import BaseProjectConfig
 
+# inherit from SfdxOrgConfig
+from cumulusci.core.config.scratch_org_config import ScratchOrgConfig
+
 # inherit from OrgConfig
 from cumulusci.core.config.sfdx_org_config import SfdxOrgConfig
 
-# inherit from SfdxOrgConfig
-from cumulusci.core.config.ScratchOrgConfig import ScratchOrgConfig
-
 # inherit from BaseProjectConfig
 from cumulusci.core.config.universal_config import UniversalConfig
-
 
 __all__ = (
     "FAILED_TO_CREATE_SCRATCH_ORG",
